requirements.txt
ttkbootstrap
python-docx
